ajouter d'une page pour reset la base et ajout des fonctions associées

+++ migrations/2024-04-30-084754_create_positions_history/up.sql
@@ -4,7 +4,7 @@ CREATE TABLE positions_history (
     latitude FLOAT NOT NULL,
     longitude FLOAT NOT NULL,
     protected_id INTEGER NOT NULL,
-    timestamp INTEGER NOT NULL,
+    timestamp BIGINT NOT NULL,
     PRIMARY KEY (protected_id, timestamp),
     FOREIGN KEY (protected_id) REFERENCES protected(id)
 )
